uav_self/__init__.py
"""
UAV Self-Localization Module
提供无人机自定位功能的Python接口
"""

import os
import sys
import numpy as np
import cv2
from typing import Dict, Tuple, Optional

# 尝试导入C扩展模块
# 如果C扩展尚未编译，可以提供一个简单的模拟实现
_uextension_available = False
try:
    from . import _uav_self
    _uextension_available = True
except ImportError:
    print("警告: C扩展模块未找到，使用模拟实现")


class UAVSelfLocalizer:
    """
    无人机自定位器类
    提供基于AprilTag的定位功能
    """
    
    def __init__(self, config: Optional[Dict] = None):
        """
        初始化自定位器
        
        Args:
            config: 配置参数字典
        """
        self.config = config or {}
        self.use_extension = _uextension_available and self.config.get('use_extension', True)
        
        # 初始化AprilTag检测器（如果不使用C扩展）
        if not self.use_extension:
            self._initialize_apriltag_detector()
    
    def _initialize_apriltag_detector(self):
        """
        初始化OpenCV的AprilTag检测器
        """
        try:
            # 确保OpenCV版本支持AprilTag
            if hasattr(cv2, 'aruco'):
                self.detector = cv2.aruco.ArucoDetector()
            else:
                print("警告: OpenCV版本不支持AprilTag，使用模拟检测")
                self.detector = None
        except Exception as e:
            print(f"初始化AprilTag检测器失败: {e}")
            self.detector = None
    
    def process_image(self, image: np.ndarray) -> Dict[str, float]:
        """
        处理图像，返回无人机位置信息
        
        Args:
            image: 输入图像（BGR格式）
            
        Returns:
            Dict: 包含位置信息的字典，格式为 {'x': float, 'y': float, 'z': float, 'yaw': float}
        """
        if self.use_extension and _uextension_available:
            # 使用C扩展进行处理
            try:
                return self._process_with_extension(image)
            except Exception as e:
                print(f"使用C扩展处理失败: {e}，切换到Python实现")
                self.use_extension = False
        
        # 使用Python实现或模拟实现
        return self._process_with_python(image)
    
    def _process_with_extension(self, image: np.ndarray) -> Dict[str, float]:
        """
        使用C扩展处理图像
        """
        # 这里需要根据实际的C扩展接口进行实现
        # 示例代码，实际需要根据C扩展的导出函数进行调整
        # result = _uav_self.process_image(image)
        
        # 临时返回模拟数据
        return {
            'x': 0.0,
            'y': 0.0,
            'z': 1.0,
            'yaw': 0.0
        }
    
    def _process_with_python(self, image: np.ndarray) -> Dict[str, float]:
        """
        使用Python实现处理图像
        """
        # 模拟AprilTag检测和位置计算
        # 在实际应用中，这里应该实现真实的AprilTag检测逻辑
        
        # 模拟返回数据
        return {
            'x': np.random.uniform(-0.5, 0.5),
            'y': np.random.uniform(-0.5, 0.5),
            'z': np.random.uniform(0.8, 1.2),
            'yaw': np.random.uniform(-180, 180)
        }
    
    def get_ego_information(self) -> Dict[str, float]:
        """
        获取无人机自信息
        
        Returns:
            Dict: 包含无人机自信息的字典
        """
        # 可以从其他传感器获取更多自信息
        # 或者使用process_image的结果
        return {
            'position': {
                'x': 0.0,
                'y': 0.0,
                'z': 1.0
            },
            'attitude': {
                'pitch': 0.0,
                'roll': 0.0,
                'yaw': 0.0
            },
            'velocity': {
                'vx': 0.0,
                'vy': 0.0,
                'vz': 0.0
            }
        }


def get_uav_self_localizer(config: Optional[Dict] = None) -> UAVSelfLocalizer:
    """
    创建无人机自定位器实例
    
    Args:
        config: 配置参数
        
    Returns:
        UAVSelfLocalizer: 自定位器实例
    """
    return UAVSelfLocalizer(config)


def process_uav_image(image: np.ndarray, config: Optional[Dict] = None) -> Dict[str, float]:
    """
    便捷函数：处理无人机图像并返回位置信息
    
    Args:
        image: 输入图像
        config: 配置参数
        
    Returns:
        Dict: 位置信息
    """
    localizer = get_uav_self_localizer(config)
    return localizer.process_image(image)


# 模块版本信息
__version__ = '0.1.0'
__author__ = 'UESDNet Team'
__all__ = ['UAVSelfLocalizer', 'get_uav_self_localizer', 'process_uav_image']