requirements.txt
# PyTorch 核心库
pytorch==1.13.1
torchvision==0.14.1
cudatoolkit==11.7

# 数值计算和科学计算
numpy==1.23.5
scipy==1.10.0

# 配置文件处理
pyyaml==6.0
omegaconf==2.3.0

# 图像处理
opencv-python==4.7.0.72
pillow==9.4.0
imageio==2.25.1

# 数据增强
albumentations==1.3.0
imgaug==0.4.0

# 进度条和日志
tqdm==4.65.0
loguru==0.7.0

# 评估指标
scikit-learn==1.2.2
pycocotools==2.0.6

# 可视化
matplotlib==3.7.1
seaborn==0.12.2

# 实验跟踪
wandb==0.14.0
tensorboard==2.12.3

# 模型导出
onnx==1.13.1
onnxruntime==1.14.1

# 开发工具
pytest==7.3.0
pytest-cov==4.1.0
flake8==6.0.0
black==23.3.0
isort==5.12.0

# 其他工具
# C扩展支持
setuptools>=65.5.0
wheel>=0.38.4

# UAV自定位模块依赖
numpy>=1.23.5
opencv-python>=4.7.0.72
gdown==4.7.1  # 用于从Google Drive下载文件
requests==2.28.2
pandas==1.5.3
jupyter==1.0.0  # 用于开发和调试